# backend/src/rest/v1/openapi.yaml
openapi: 3.0.3
info:
  title: ChatFlow API
  description: A minimal but clean WhatsApp-style messaging app API
  version: 1.0.0
  contact:
    name: ChatFlow Team
servers:
  - url: http://localhost:3002/v1
    description: Development server
  - url: https://chatflow-backend-3w6u4kmniq-ue.a.run.app/v1
    description: Production server (Cloud Run)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          pattern: '^conv_[0-9]+_[a-z0-9]+$'
          description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
          example: 'conv_1750386041311_fpmswok2p'
        senderId:
          type: string
          format: email
        messageType:
          type: string
          enum: [TEXT, IMAGE, FILE]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResult:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
    CreateMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          oneOf:
            - type: string
              minLength: 1
              description: "Plain text content (not recommended for production)"
              example: "Hello, this is a plain text message"
            - $ref: '#/components/schemas/EncryptedField'
          description: "Message content - should be encrypted for privacy"
        messageType:
          type: string
          enum: [TEXT, IMAGE, FILE]
          default: TEXT
          description: "Type of message (unencrypted for routing purposes)"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details: {}
...
    UpdateMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 10000
              description: "Plain text content (not recommended for production)"
              example: "Updated plain text message"
            - $ref: '#/components/schemas/EncryptedField'
          description: "Updated message content - should be encrypted for privacy"
...
    EncryptionMetadata:
      type: object
      description: "Metadata for field-level encryption using AES-256-GCM"
      required:
        - algorithm
        - keyId
        - iv
        - tag
      properties:
        algorithm:
          type: string
          enum: ["AES-256-GCM"]
          description: "Encryption algorithm used"
          example: "AES-256-GCM"
        keyId:
          type: string
          description: "Client-side key identifier for key management"
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "user_key_001"
        iv:
          type: string
          format: base64
          description: "Base64 encoded initialization vector (96-bit for GCM)"
          example: "MTIzNDU2Nzg5MGFiY2RlZg=="
        tag:
          type: string
          format: base64
          description: "Base64 encoded authentication tag (128-bit for GCM)"
          example: "dGFnMTIzNDU2Nzg5MGFiY2Q="
        timestamp:
          type: integer
          format: int64
          description: "Unix timestamp for replay attack prevention"
          example: 1640995200
        nonce:
          type: string
          description: "Unique nonce for each encryption operation"
          example: "unique_nonce_001"
    EncryptedField:
      type: object
      description: "Encrypted field with metadata"
      required:
        - data
        - encryption
      properties:
        data:
          type: string
          format: base64
          description: "Base64 encoded encrypted data"
          example: "eyJhbGciOiJBMjU2R0NNIiwiZW5jIjoiQTI1NkdDTSJ9..."
        encryption:
          $ref: '#/components/schemas/EncryptionMetadata'
paths:
...
  /conversations/{conversationId}/messages:
    get:
      tags:
        - Messages
      summary: Get messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Messages retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginationResult'
                          - type: object
                            properties:
                              data:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Message'

    post:
      tags:
        - Messages
      summary: Send a message to a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
  /conversations/{conversationId}/messages/{messageId}:
    get:
      tags:
        - Messages
      summary: Get a specific message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            pattern: '^msg_[0-9]+_[a-z0-9]+$'
            description: 'Message ID in format: msg_{timestamp}_{randomString}'
            example: 'msg_1750386041311_abc123def'
      responses:
        '200':
          description: Message retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Messages
      summary: Update a specific message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            pattern: '^msg_[0-9]+_[a-z0-9]+$'
            description: 'Message ID in format: msg_{timestamp}_{randomString}'
            example: 'msg_1750386041311_abc123def'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMessageRequest'
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Messages
      summary: Delete a specific message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: '^conv_[0-9]+_[a-z0-9]+$'
            description: 'Conversation ID in format: conv_{timestamp}_{randomString}'
            example: 'conv_1750386041311_fpmswok2p'
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            pattern: '^msg_[0-9]+_[a-z0-9]+$'
            description: 'Message ID in format: msg_{timestamp}_{randomString}'
            example: 'msg_1750386041311_abc123def'
      responses:
        '204':
          description: Message deleted successfully
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...